requirements.txt
asgiref==3.2.7
Django==3.0.5
django-widget-tweaks==1.4.8
pytz==2020.1
sqlparse==0.3.1
razorpay==1.3.0
django-tempus-dominus==5.1.2.17